refactor: change dose log amount to float

+++ app/src/main/java/com/example/pillmate/data/local/entity/SupplyLogEntity.kt
@@ -8,7 +8,7 @@ data class SupplyLogEntity(
     @PrimaryKey
     val id: String, 
     val medicationId: String, 
-    val changeAmount: Int,
+    val changeAmount: Float,
     val reason: String,
     val timestamp: Long 
 )
